+++ src/utils/axios-films.js
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const instanceFilms = axios.create({
+  baseURL: 'https://swapi.co/api/films/'
+});
+
+export default instanceFilms;
